Add files via upload
I created this Tic-Tac-Toe game following the guidelines of rushoni02. This game was built using HTML, CSS, and JavaScript.

The game mode brings the traditional format of Tic-Tac-Toe. There is a button to restart the game. When the game is restarted, it gives the second player the opportunity to start the turn, thus alternating who starts each game.

I hope this helps. Best regards!